backend/requirements.txt
django>=5.0,<6
djangorestframework>=3.15
django-cors-headers>=4.3
python-dotenv>=1.0
openpyxl>=3.1
requests>=2.31
python-docx>=1.1
reportlab>=4.0
matplotlib>=3.8
